requirements.txt
anyio==3.7.0
async-timeout==4.0.2
bcrypt==4.0.1
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
email-validator==2.0.0.post2
fastapi==0.95.2
h11==0.14.0
hiredis==2.2.3
idna==3.4
passlib==1.7.4
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
pymongo==4.3.3
pyotp==2.8.0
pypng==0.20220715.0
python-dotenv==1.0.0
python-jose==3.3.0
redis==4.5.5
rsa==4.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.2
uvicorn==0.22.0
